AL-3900: rich message factory pattern refactor
