+++ kommunicateui/src/main/java/com/applozic/mobicomkit/uiwidgets/conversation/richmessaging/adapters/AlListRMAdapter.java
@@ -21,19 +21,14 @@ import com.bumptech.glide.Glide;
 import java.util.List;
 import java.util.Map;
 
-public class AlRichListsAdapter extends RecyclerView.Adapter {
+public class AlListRMAdapter extends ALRichMessageAdapter {
 
-    private Context context;
     private List<ALRichMessageModel.AlElementModel> elementList;
-    private ALRichMessageListener messageListener;
-    private Message message;
     private Map<String, Object> replyMetadata;
 
-    public AlRichListsAdapter(Context context, Message message, List<ALRichMessageModel.AlElementModel> elementList, Map<String, Object> replyMetadata, ALRichMessageListener messageListener) {
-        this.context = context;
+    AlListRMAdapter(Context context, Message message, List<ALRichMessageModel.AlElementModel> elementList, Map<String, Object> replyMetadata, ALRichMessageListener messageListener) {
+        super(context, messageListener, message);
         this.elementList = elementList;
-        this.messageListener = messageListener;
-        this.message = message;
         this.replyMetadata = replyMetadata;
     }
 
@@ -46,10 +41,14 @@ public class AlRichListsAdapter extends RecyclerView.Adapter {
 
     @Override
     public void onBindViewHolder(@NonNull RecyclerView.ViewHolder holder, int position) {
-        bindView((AlListItemViewHolder) holder, elementList.get(position));
+        bindItems(holder, position);
     }
 
-    public void bindView(AlListItemViewHolder holder, ALRichMessageModel.AlElementModel element) {
+    @Override
+    public void bindItems(RecyclerView.ViewHolder viewHolder, int position) {
+        ALRichMessageModel.AlElementModel element = elementList.get(position);
+        AlListItemViewHolder holder = (AlListItemViewHolder) viewHolder;
+
         if (!TextUtils.isEmpty(element.getTitle())) {
             holder.headerTv.setVisibility(View.VISIBLE);
             holder.headerTv.setText(AlRichMessage.getHtmlText(element.getTitle().trim()));
@@ -101,8 +100,8 @@ public class AlRichListsAdapter extends RecyclerView.Adapter {
             if (itemPosition != -1 && elementList != null && !elementList.isEmpty()) {
                 if (context.getApplicationContext() instanceof ALRichMessageListener) {
                     ((ALRichMessageListener) context.getApplicationContext()).onAction(context, getAction(elementList.get(itemPosition)), message, elementList.get(itemPosition), getReplyMetadata(elementList.get(itemPosition)));
-                } else if (messageListener != null) {
-                    messageListener.onAction(context, getAction(elementList.get(itemPosition)), message, elementList.get(itemPosition), getReplyMetadata(elementList.get(itemPosition)));
+                } else if (listener != null) {
+                    listener.onAction(context, getAction(elementList.get(itemPosition)), message, elementList.get(itemPosition), getReplyMetadata(elementList.get(itemPosition)));
                 }
             }
         }
